Port TestDeleteNodeWiki
